requirements.txt
astroid==2.2.5
autopep8==1.4.3
Django==2.1.7
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-mptt==0.9.1
factory-boy==2.11.1
Faker==1.0.4
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
text-unidecode==1.2
typed-ast==1.3.1
wrapt==1.11.1
